Made some more fixes to example files.
Still haven't found the problems with rendering speed, but my guess is that it has something
to do with the fact that we're not drawing directly to the root window. For now, I'll consider
this release to be an experimental one, not a proper one. It doesn't do anything new, anyway,
other than provide the turtle canvas as a component instead of a dedicated window.